pyproject.toml
[project]
name = "mb-pomodoro"
version = "0.0.3"
description = "macOS Pomodoro timer with a CLI-first workflow"
readme = "README.md"
license = "MIT"
authors = [{ name = "mcbarinov" }]
requires-python = ">=3.14"
dependencies = ["typer~=0.24.0", "pyobjc-framework-Cocoa~=12.1"]
[project.scripts]
mb-pomodoro = "mb_pomodoro.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest~=9.0.2",
    "pytest-xdist~=3.8.0",
    "ruff~=0.15.1",
    "pip-audit~=2.10.0",
    "bandit~=1.9.3",
    "mypy~=1.19.1",
]

[tool.mypy]
python_version = "3.14"
strict = true
exclude = ["^tests/", "^tmp/"]

[[tool.mypy.overrides]]
module = ["AppKit", "Foundation", "PyObjCTools.*", "objc"]
ignore_missing_imports = true

[tool.ruff]
line-length = 130
target-version = "py314"
[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "TC",      # flake8-type-checking, TYPE_CHECKING is dangerous, for example it doesn't work with pydantic
    "A005",    # flake8-builtins: stdlib-module-shadowing
    "ERA001",  # eradicate: commented-out-code
    "PT",      # flake8-pytest-style
    "FIX",     # flake8-fixme
    "PLR0911", # pylint: too-many-return-statements
    "PLR0912", # pylint: too-many-branches
    "PLR0913", # pylint: too-many-arguments
    "PLR2004", # pylint: magic-value-comparison
    "PLC0414", # pylint: useless-import-alias
    "FBT",     # flake8-boolean-trap
    "EM",      # flake8-errmsg
    "TRY003",  # tryceratops: raise-vanilla-args
    "C901",    # mccabe: complex-structure,
    "BLE001",  # flake8-blind-except
    "S311",    # bandit: suspicious-non-cryptographic-random-usage
    "TD002",   # flake8-todos: missing-todo-author
    "TD003",   # flake8-todos: missing-todo-link
    "RET503",  # flake8-return: implicit-return
    "COM812",  # it's used in ruff formatter
    "D203",    # pydocstyle: one-blank-line-before-class (conflicts with D211)
    "D213",    # pydocstyle: multi-line-summary-second-line (conflicts with D212)
]
[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["field_validator"]
[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ANN", "S", "D104"]
[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B311"]
